refactor: apply some mechanical cleanups
- Move Python preamble parsing out of VirtualWriter into py_preamble.rs
- Extract require_single_interpreter for PEP 517 hooks
- Use host Target venv helpers in PGO instrumentation
- Remove SBOM pass-through wrappers on BuildOrchestrator
- Make ArtifactContext.sbom a plain SbomConfig
- Remove lossy Display for StableAbi; use stable_abi.kind at call sites
- Fix tags_from_bridge panic on empty interpreter list
- Deduplicate the missing-cdylib error message
- Doc/import/message hygiene fixes in bridge and build_orchestrator
