+++ deny.toml
@@ -77,8 +77,6 @@ feature-depth = 1
 ignore = [
     # paste unmaintained
     "RUSTSEC-2024-0436",
-    # num_prefix unmaintained
-    "RUSTSEC-2025-0119",
     #{ id = "RUSTSEC-0000-0000", reason = "you can specify a reason the advisory is ignored" },
     #"a-crate-that-is-yanked@0.1.1", # you can also ignore yanked crate versions if you wish
     #{ crate = "a-crate-that-is-yanked@0.1.1", reason = "you can specify why you are ignoring the yanked crate" },
